requirements.txt
schedule==1.2.0
crawl4ai==0.7.2
langchain==0.3.27
langchain-community==0.3.27


langgraph==0.6.2
openai>=1.68.2

requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.11.7
beautifulsoup4==4.12.2
newspaper3k==0.2.8
feedparser==6.0.10
python-dateutil==2.8.2
markdown2==2.4.10
